requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==4.2.16
django-appconf==1.0.6
django-colorfield==0.11.0
django-compressor==4.5.1
django-extensions==3.2.3
django-libsass==0.9
django-netfields==1.3.2
django-postgres-backup==1.0.0
djangorestframework==3.15.2
idna==3.10
iniconfig==2.0.0
libsass==0.23.0
netaddr==1.3.0
packaging==24.2
paramiko==3.5.0
pillow==11.0.0
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyNaCl==1.5.0
pytest==7.4.4
rcssmin==1.1.2
requests==2.32.3
rjsmin==1.2.2
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3
